Datacore Maintenance (#5099)
# About the pull request
A bit of maintenance for the generic datacore, changes one letter vars &
makes it so that when viewing the list of datacore entries for human
players it is possible to determine who's entry is who's without opening
up a hundred of them one by one.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
QOL
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: Cleaned up 1 letter vars in the global datacore. Also made the
entry names reflect the real name of who they relate to, for VV
purposes.
/:cl: