Paygrade shorthand defines. (#5128)
# About the pull request
This PR converts all paygrade shorthands (e.g. ME1) into PAY_SHORT_XXX
defines. The reason for this is to ensure any changes are reflected
across the board, and also to allow the shorthand to carry description
of what it is. It can be hard to tell what a paygrade is meant to be
based upon arbitrary letters.

# Explain why it's good for the game

Makes future code changes easier.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: Paygrade shorthands now use defines.
fix: CLF and Survivors are no longer privates.
fix: Civillian Mr/Ms/Mx prefix now works correctly, and is the default
for new ID cards.
/:cl: