Fixes disarm chance calculation (#5111)
# About the pull request

I was testing stuff out for a different PR and thought the disarm chance
(for human vs human) seemed weird at times, so I forced it to output
attacker_skill_level and defender_skill_level to chat every time a
disarm was done. And then found out it's not calculating it properly.
This seems to fix it in limited testing but might need more testing. I'm
kinda surprised it wasn't noticed earlier cause there's two signs it's
wrong in certain situations even with skill_level variables being
invisible: the "tackled" message should only appear when the attacker's
CQC level is two or higher (otherwise it should say pushed/shoved), and
also, anyone that has a CQC advantage is still disarm-stunning everyone
for the same amount of time as anyone else on average ('on average'
because there was apparently some variance with a tick offset bug before
the TG status effect testmerge) - the code implies it should be longer.

# Explain why it's good for the game

Bugs should be fixed. From a balance perspective I don't know if it
'should' be fixed from what it is currently in it's bugged state... but
regardless, it's a bug.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: disarm chance wasn't calculating properly
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>