Ckey check verb (#5073)
# About the pull request
Stops requirement for proccall on handy feature for staff.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
admin: Added a Check Ckey verb to analyze for multikey. We're watching
you Wazowski. Always watching.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>