Added launch alarm for PR #4976 (#5105)
# About the pull request

Adds a looping alarm sound when the lifeboats are launched. This also
uses the function described in PR #4976 to trigger the sound. This is
NOT for adding those features, so wait until that is merged to merge
this PR.

# Explain why it's good for the game

Denies the need to make an announcement warning marines that the
lifeboats are leaving.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://youtu.be/L8ia2OKhUUA



</details>


# Changelog
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't



match. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:alexguinea atticus-rezzer
add: Adds alarm sound for lifeboats when launching
/:cl:

---------

Co-authored-by: zzzmike <85382350+zzzmike@users.noreply.github.com>