refactor shutter and linked buttons in CL office. (#5149)
# About the pull request
basically refactoring shutter and buttons also adding lights because it
was too dark for an office and some button and separation for qol while
i was there...
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
1-i added the three door because it's dark
2-i separated the shutter for the windows and the door because it seem
like it's been something simple that make sense
3-added the button to the shutter for the door inside because it's way
more practical.
4-refactoring code to be less dmm dependent because it's cleaner and
avoid to touch dmm file if you want to modify them in the future...
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: added a button inside the CL office to open the shutter for
the door and rearranged other buttons.
maptweak: separated shutter for cl office door and for the windows.
maptweak: add three lights in CL office and quarter.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>