Balloon Alerts for Gun Safety (#5152)
# About the pull request
When attempting to shoot a gun while safety is on (either the gun safety
itself or triggering help intent safety) you'll get a balloon alert
informing you.
help intent safety is on (will be in lowercase):

![image](https://github.com/cmss13-devs/cmss13/assets/54422837/1942e2ae-546b-4ab3-814f-2ef0e425544d)
gun safety toggled on:

![image](https://github.com/cmss13-devs/cmss13/assets/54422837/631db7b8-c266-42f5-a322-6d80882b86e1)

also decreases the size of the "Help intent safety is on! Switch to
another intent to fire your weapon." message to normal size.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Reading chat when you're (presumably) in the middle of trying to shoot
at a xeno is quite hard, having it as a balloon alert allows you to
quickly realize what's happening.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Gun safety messages now use balloon alerts.
/:cl: