PO can remove queens override again, bit by bit edition (#5143)
# About the pull request
PO can perform a 20 seconds do_after to remove additional 20 seconds of
the lockdown.

# Explain why it's good for the game
You used to be able to remove lockdown completely if you perform a 3
minutes do_after, but it was lost during the port to new shuttle system.
Now it's back in another, better form! Gives marines a chance to remove
lockdown sooner if they control the cockpit. In best case scenario you
can remove the lock in 5 minutes.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
balance: PO can remove queens dropship override again.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>