Minor sound changes and clarity. (#5088)
# About the pull request

Adds the armorequip noise when backpack, uniform, shoes are unequipped.
Changes the whoosh (disarm-stun) noise to only trigger when you stun
someone. Adds sound to tablestun, and changes tablestun text to be more
clear to new players that it stuns.

# Explain why it's good for the game

In human vs human fights, sometimes your backpack/uniform/shoes get
stolen. The armorequip sound is a good clue to draw attention to when
this happens, both for the victim and the attacker, so adding it to a
few more important gear pieces should be a benefit in these fringe
cases. The effect is subtle enough, and the situation rare enough, that
I don't think it would be too much. The whoosh sound is currently
triggering when you "disarm" someone even when they are holding nothing
in hand, which doesn't make sense. Now it's reserved for when a
disarm-stun goes off. Swoosh sound added to tablestun, because tablestun
being silent doesn't make sense. From my experience talking to players,
tablestun is not very well known outside of veteran players who
familiarized themselves with that niche interaction, so adding a sound
and changing the text to make the effect clear should be a benefit.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

(https://www.youtube.com/watch?v=ElOAaq0dYcg&ab_channel=MK)

</details>


# Changelog
:cl:
spellcheck: tablestun text is more clear that it stuns
soundadd: unequip noise added to uniform, backpack, shoes. 
soundadd: sound added to tablestun
sounddel: for disarms, whoosh only triggers on stun. otherwise, the miss
sound plays
/:cl: