Fusion reactor explosion handling (#5166)
# About the pull request
You cannot destroy fusion reactors with explosives. Explosions do
disable overloading tho as a little feature.

# Explain why it's good for the game
Not only it causes bugs, reactors is an important objective and you
should not be able to destroy them with HEDP.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
fix: fusion reactors cannot be destroyed with explosions
/:cl: