renamed 'M56 Battery' to 'M56 DV9 Battery' in requisitions munition vendor (#5196)
# About the pull request
Renames 'M56 Battery' in requisitions munition vendor to 'M56 DV9
Battery'

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
A Smartgunner will ask a Cargo Tech for a DV9 and they will be unable to
search for one causing confusion. The battery itself is called a DV9
which is what the SG asks for and this edit fixes the vendor not listing
it as such.
# Testing Photographs and Procedure
Minor text change. I verified it in-game and it's working correctly.


# Changelog
:cl:
spellcheck: Edited 'M56 Battery' to 'M56 DV9 Battery' in requisitions
munition vendor
/:cl: