Compass Headgear now shows your direction when examined (#5198)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
When the compass headgear is now examined, it will tell you the
direction your character is facing (north, south, east, or west). This
only works when you're on a planet, this means being groundside and not
on a map that is situated in space.

I've considered changing the description, but I've left it because I
couldn't think of a better one.

# Explain why it's good for the game
When I used the compass in game, I expected to actually tell me which
direction I'm facing, but I was a bit disappointed when it didn't. So, I
change it to work in a realistic way.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

On LV-624, where the compass works (left). On the ship, were the compass
does not work (right).

![image](https://github.com/cmss13-devs/cmss13/assets/91113370/323a14ce-551f-4d38-8a25-a3b666f98226)

On maps like Fiorina where are situated in space, the compass will not
work.

![image](https://github.com/cmss13-devs/cmss13/assets/91113370/4548efd9-c674-400e-af38-0bfb4c7749c8)

</details>


# Changelog
:cl:
add: The compass headgear will now tell you what direction you're facing
when on a planet.
code: Added a new ground environmental trait for a ground level in
space.
/:cl:
My discord is: contrabang, if you would like to contact me, that is the
best place.