Fix plane when dead nested mobs merge with weeds (#5180)
# About the pull request

This PR fixes an issue where dead nested mobs get set to the floor plane
thus disappearing under the wall. The signal for nest deletion is simply
so we can set the plane to the floor plane when dropped to the ground
since it is intentional weeded mobs get set to the floor plane so they
layer under more things.

# Explain why it's good for the game

Visually corrects mobs that died while in a nest.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/c1370f51-05c8-4f43-b938-0e1ff6bfce2f)

</details>


# Changelog
:cl: Drathek
fix: Fix dead nested mobs disappearing when weeded
/:cl: