Xeno telegraph effect can now be any rgb colour (#5207)
# About the pull request

I didn't want it to be locked into the basic colours we have based on
the sprites, so instead you can supply a colour code to the constructor
to create any colour you want. Defaults to white if none provided.
Eventually would be good to roll this out also the animated sprites.

# Explain why it's good for the game

When we add new castes/xenos we can diversify the colour palette for
different effects instead of everything being red or brown.


https://github.com/cmss13-devs/cmss13/assets/25027759/be086245-eda3-4f27-8ce0-525df59674a1

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Xenomorph telegraph effects can now be any rgb colour.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>