Refactor got handling infection stage and progess of the black goo disease (#5044)
# About the pull request

1- reduce the number of stages from 5 to 3. each stage take 6 min or
360s

2-rework the logic that handle the disease progress

A-Set stage_prob to zero to prevent random stage increase.

B-add a stage_level variable that when is at-least 360 make an increase
of the disease stage...

C-add an infection_rate variable that handle how fast the stage_level
increase passively.
( stage_level = stage_level +infection_rate )
i also made it so that infection-rate is quadruple when the host is
dead.

D-I also added that the goo messages (you feel warm ....) give a boost
to stage_level to add some randomness to the infection progress.

To do list:
i need to figure out how to increase stage level when someone get hit by
a zombie after he is infected...
no idea if i can manage that so i am just going to put what i made up
for review...

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
those change are to make infection evolution less random and way more
smooth progress with the addition of some random boost with the good
message since it's something that could represent a natural boost of the
infection progress in the host.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: reduce the number of stages from 5 to 3. each stage take 6 min
or 360s.
balance: rework the logic that handle the disease progress
balance: A-Set stage_prob to zero to prevent random stage increase.
balance: B-add a stage_level variable that when is at-least 360 make an
increase of the disease stage...
balance: C-add an infection_rate variable that handle how fast the
stage_level increase passively.
balance: D-infection_rate is quadruple when the host is dead.
balance: E-I also added that the goo messages (you feel warm ....) give
a boost to stage_level to add some randomness to the infection progress.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>