Standartize the price of all three binculars type for SG,medic,comtech, IO (#5188)
# About the pull request
The logic for standartizing it go as fallow:
no reason to not make lower tier of bino available if you have access to
higher tier of it.
no reason to not give access to higher tier if lower ranks have access
to it.

concerning the price change for medic i found it weird that it was the
only one with this price tag so i figure better smooth things up. and
their is no reason for medic to need higher tier binocular as they role
isn't to do JTAC...

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Give access to more items while keeping it logical 
Giving people a cheaper option than higher tier one
also those bino are free in the prep vendor and have a few public spot
where they spawn.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: For SG,IO,medic,comtech standartize the binoculars section as
fallow (Binoculars=5, Range Finder =10, Laser Designator=15).
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>