fix bug with landing zone camera. (#5195)
# About the pull request
basicly a bunch of camera badly place would get attach to DS when DS
land on them meaning they could no longer function as intented.

fixes: https://github.com/cmss13-devs/cmss13/issues/5071
some camera where badly placed so it decided to standardize all the
camera placement of the map in rotation?


<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: added landing zone camera on trijent.
fix: fixed landing zone camera on all map in rotation.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>