Separates Factions In Observe/Follow Tab (#5058)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Separates the factions from a generic 'ERT Faction' tab into their own
separate one's for each of the main factions, this includes.

- WY
- UPP
- CLF
- Freelancers
- Mercenaries
- Contractors
- Dutch's Dozen
- TWE RMC
- CMB Marshals

The remaining factions (mostly memes, unused or admin ones) are kept in
the current 'ERT' panel.

# Explain why it's good for the game

Better gauge as a observer who the present factions are, helps in the
uncommon circumstances where multiple factions are in play.


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl:
add: Factions are now properly separated in the observe menu as a ghost.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>