Fix Santa Hats breaking due to map camouflage (#5268)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
Simple as it can be - the map camouflage was applied on santa helmets
when they don't have camouflage variants

# Explain why it's good for the game
Spirit of holidays and all that

# Testing Photographs and Procedure
Spawn on a camo-enabled map


# Changelog
:cl:
fix: Fix Xmas helmets getting overriden by map camouflage.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>