Ghost listening device preference (#5219)
# About the pull request
Ghosts can now toggle between hearing and not hearing listening devices.

## THIS DEFAULTS TO OFF

# Explain why it's good for the game
Bit of peace of mind, a lot of text can be duplicated by hearing
listening devices as a ghost.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Ghosts can now toggle hearing listening devices or not. This is
dependant on hearing radios and will not function to allow hearing
devices without also hearing radios.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>