Fix Xmas barricade decorator not working (#5270)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
Decorator changed the main icon of barricades. Not only this is bad due
to the file not being up to date, but also wiring doesn't care and
reuses the builtin icon file.

# Explain why it's good for the game
Event code not working

# Testing Photographs and Procedure
Spawn in, test


# Changelog
:cl:
fix: Fixed X-mas barricade wiring not applying properly.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>