Allows Xenomorphs to turn off the dropship's launch alarm (#5210)
# About the pull request

Allows Xenomorphs to turn off the dropship's launch alarm without the
queen needing to launch.

# Explain why it's good for the game

Occasionally (2-3 times that I can think of, at least) someone turns on
the dropship's launch alarm after the queen calls it down, resulting in
a situation where if the queen wants to wait for evolutions or something
else, the entire hive has to listen to a beeping noise for multiple
minutes.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/57483089/9014b2bb-a68a-4ef8-b7ee-1f3e03ce2575

</details>


# Changelog
:cl:
add: Allowed Xenomorphs to turn off the dropship's launch alarm by
hitting the nav computer.
/:cl: