New Healer Drone Ability: Sacrifice, another go (#5173)
# About the pull request

New healer drone skill — sacrifice. If things get really bad, the healer
can instantly transfer 75% of its current health, but for the great cost
of the healer’s own life.

The sacrifice ability is an emergency button. You don’t have to use it,
but at some point you may find that sacrifice is a right thing to do, so
it’s a nice option to have. I want to emphasize that this skill is not a
necessary part of the healer gameplay. You don’t have to trade your life
for the life of another xeno. But you can if you are willing to.

Also healer drones gets a counter of the health it has transferred. If
healer drone transfers at least 10 000 health points, it will be able to
respawn after using the sacrifice ability (like acid runner after
kaboom).

# Explain why it's good for the game

Saving fellow xenos is the most satisfying part of healer drone’s
gameplay. This ability allows to save others in really dire situations,
but for the great cost. It’s still fitting, because drone already can
use heal ability until it goes into crit or dies, in some situations
it’s a worthy risk. Also rewards long living healer drones with the
opportunity to use the ability without dying.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
balance: New healer drone ability — sacrifice. Healer drone can
instantly transfer 75% of its current health, but for the great cost of
the healer’s own life.
/:cl:

---------

Co-authored-by: ihatethisengine <treml.treml@yandex.ru>
Co-authored-by: harryob <me@harryob.live>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>