Executive Officer's Weapons Vendor + More Belt Options (#5179)
# About the pull request

This PR adds a new 'weapons' vendor for the Executive Officer to acquire
a suite of personal gear for their use. It also adds a new belt category
to their uniform vendor to give the Executive Officer more options in
what kind of belt they want to take.

The XO can chose to vend either a MK1 M41A weapon's kit, equilivant to
the armoury, or a MK221 shotgun kit which comes with a box of buckshot
and slug rounds each.

# Explain why it's good for the game

The Executive Officer exists as a sort of stepping stone to the
Commanding Officer to a degree, because of this, it has been somewhat
languishing in-between the Commander and Staff Officers.

The CO has a full suite of personal vendors and unique equipment to pick
from, and the Staff Officers have their own unique weapons vendor and
ancillary equipment from the armoury they can select, including special
medical and engineering kits.

The Executive Officer only has a uniform vendor to draw from, but
otherwise they are forced to skim off of the armoury and the staff
officer's equipment.

Due to the recent rule changes that allow the XO to deploy more freely
groundside, as well as deploying during a Nuclear weapon deployment, I
feel the time is now to address the issue.

Ergo, this PR adds a new vendor to give the XO more options on what
equipment they can take for when they deploy, or for hijack. The
equipment is somewhat in keeping with what the CO gets, minus the
special equipment. Ergo, it can help an aspiring CO get a better idea of
how their setup works for the CO just as it does for a XO.

In terms of abuse, I do not believe this will be a issue. The XO is held
to the highest standards of roleplay outside of whitelisted roles, an XO
already has the power to loot the armoury or requisitions, I do not
believe this will enable something negative considering they already
have that ability if they want to.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2023-12-10 16 24
09](https://github.com/cmss13-devs/cmss13/assets/6595389/67eea32f-8192-406f-9684-fe70e22b6542)

![Screenshot 2023-12-10 16 24
02](https://github.com/cmss13-devs/cmss13/assets/6595389/74cb5419-c34d-4c53-beb9-eb3e57fdb4e2)


</details>


# Changelog
:cl:
add: The Executive Officer now has a personal weapons vendor. It
includes the ability to acquire a full suite of combat and support gear,
giving the XO more agency in customizing their loadout.
add: Adds a new set of belts to the Executive Officer's uniform vendor. 
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>